From SqlServer-2017 to Mysql-8.0.22

+++ admin/src/main/java/com/vincent/admin/vo/Pagination.java
@@ -10,8 +10,8 @@ import lombok.Data;
 @Data
 public class Pagination<T> {
 
-    private Long currentPage;
+    private Integer currentPage;
 
     // 分页大小
-    private Long pageSize;
+    private Integer pageSize;
 }
